% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_hamming}
\alias{compute_hamming}
\title{Compute Hamming Distance}
\usage{
compute_hamming(z, model_data)
}
\arguments{
\item{z}{Integer vector. The sequence data, flattened (individuals x time points).}

\item{model_data}{List. The model data object, containing \code{n_time} and \code{n_id}.}
}
\value{
A \code{dist} object containing the pairwise Hamming distances between sequences.
}
\description{
Internal: Computes the Hamming distance between sequences represented in a vector \code{z}.
}
\keyword{internal}
