% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_data.R
\name{create_model_data}
\alias{create_model_data}
\title{Create model data}
\usage{
create_model_data(
  z,
  id,
  time,
  n_basis = 10,
  intercept_penalty = 1,
  dirichlet_param = 1
)
}
\description{
...
}
