% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_model_data_min}
\alias{create_model_data_min}
\title{Create Minimal Model Data List}
\usage{
create_model_data_min(model_data, M, G)
}
\arguments{
\item{model_data}{List. The full model data object (see \code{create_model_data}).}

\item{M}{Integer. The number of clusters for the current model.}

\item{G}{Integer. The number of categories (outcome levels) for the current model.}
}
\value{
A list containing a subset of the elements from \code{model_data} (\code{M}, \code{G}, \code{w}, \code{z}, \code{x}, \code{B}, \code{P}, \code{n_vars}, \code{n_id}, \code{n_time}, \code{n}).
}
\description{
Internal: Creates a reduced list containing essential elements from the full model data.
This is used to pass a smaller object between functions when the full \code{model_data} is not required.
}
\keyword{internal}
