% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_array}
\alias{filter_array}
\title{Filter MCMC Samples}
\usage{
filter_array(array, burn_in, thin)
}
\arguments{
\item{array}{A numeric array containing MCMC samples (first dimension is iterations).}

\item{burn_in}{Integer. The number of initial iterations to discard.}

\item{thin}{Integer. The thinning rate; keep every \code{thin} iteration after burn-in.}
}
\value{
A filtered numeric array, or \code{NULL} if the input array was \code{NULL}.
The dimensions of the output array will depend on the original array's dimensions.
For a 1D array, it returns a vector. For 2D and 3D arrays, it returns a matrix or array,
respectively.
}
\description{
Internal: Filters an array of MCMC samples by applying a burn-in period and thinning rate.
}
\keyword{internal}
