% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_init_w.R
\name{find_init_overfitted}
\alias{find_init_overfitted}
\title{Find init values}
\usage{
find_init_overfitted(
  M,
  submodels,
  model_data,
  lambda,
  init_iters,
  init_thin,
  init_burn_in,
  n_cores,
  seed,
  verbose = TRUE
)
}
\description{
Find inits
}
