% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_metrics}
\alias{format_metrics}
\title{Format Model Selection Metrics}
\usage{
format_metrics(metrics_list)
}
\arguments{
\item{metrics_list}{List. Output from \code{comp_metrics}.}
}
\value{
A tibble with columns "G", "M", and the calculated metrics (l_mean, l_var, AICM, BICM).
}
\description{
Internal: Format the list of model selection metrics into a tibble.
}
