% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_basis_index}
\alias{gen_basis_index}
\title{Indices for Grouped Basis Functions}
\usage{
gen_basis_index(n_basis, M)
}
\arguments{
\item{n_basis}{Integer. Number of basis functions per group.}

\item{M}{Integer. Number of groups.}
}
\value{
List of integer vectors, one per group.
}
\description{
Creates a list of index vectors for basis functions grouped by class.
}
\examples{
\dontrun{
gen_basis_index(5, 3)
}
}
\keyword{internal}
