% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_dummy}
\alias{gen_dummy}
\title{Generate Dummy (Indicator) Matrix}
\usage{
gen_dummy(x, n_cat, intercept = FALSE)
}
\arguments{
\item{x}{Integer vector of categories (1-based).}

\item{n_cat}{Integer. Number of categories.}

\item{intercept}{Logical. If TRUE, includes an intercept column. Default is FALSE.}
}
\value{
Numeric matrix of dummy variables.
}
\description{
Creates a dummy variable (indicator) matrix for a categorical variable, with an option to include an intercept column.
}
\examples{
\dontrun{
gen_dummy(c(1, 2, 3), n_cat = 3, intercept = TRUE)
}
}
\keyword{internal}
