% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_inv_cov}
\alias{gen_inv_cov}
\title{Generate Inverse Covariance Matrices for Penalized Coefficients}
\usage{
gen_inv_cov(lambda, model_data, fixed_sd = 100)
}
\arguments{
\item{lambda}{Numeric matrix or vector of penalty parameters.}

\item{model_data}{List. Model structure (see create_model_data).}

\item{fixed_sd}{Numeric. Standard deviation for non-penalized coefficients. Default is 100.}
}
\value{
List of diagonal matrices.
}
\description{
Creates a list of diagonal inverse covariance matrices for each class (except the last).
}
\examples{
\dontrun{
gen_inv_cov(c(1, 2, 3), model_data = list(), fixed_sd = 100)
}
}
\keyword{internal}
