% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_notpen_index}
\alias{gen_notpen_index}
\title{Indices of Non-Penalized Basis Functions}
\usage{
gen_notpen_index(n_basis, M, order = 1)
}
\arguments{
\item{n_basis}{Integer. Number of basis functions per group.}

\item{M}{Integer. Number of groups or classes.}

\item{order}{Integer. Order of difference for penalization. Default is 1.}
}
\value{
Integer vector of indices.
}
\description{
Generates indices for non-penalized basis functions across groups or classes.
}
\examples{
\dontrun{
gen_notpen_index(5, 3, order = 2)
}
}
\keyword{internal}
