% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_sample_array}
\alias{gen_sample_array}
\title{Generate Initial Sample Array}
\usage{
gen_sample_array(iters, dimension, sampler = NULL, init = NULL)
}
\arguments{
\item{iters}{Integer. The total number of iterations for the MCMC chain.}

\item{dimension}{Integer vector. The dimensions of the array for each iteration's sample
(excluding the iteration dimension).}

\item{sampler}{Function. An optional function to generate initial values if \code{init} is \code{NULL}.}

\item{init}{Numeric array. Optional initial values for the first iteration.}
}
\value{
A numeric array of dimensions \code{c(iters, dimension)} with the first slice
(corresponding to the first iteration) potentially filled with \code{init}.
}
\description{
Internal: Creates an empty array to store MCMC samples, initialized with optional initial values.
}
\keyword{internal}
