% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixscat-package.R
\docType{package}
\name{mixscat}
\alias{mixscat-package}
\alias{mixscat}
\title{mixscat: Mixture-Based Statistical Clustering and Analysis Tools}
\description{
The \strong{mixscat} package provides functions for fitting and analyzing
mixture-based clustering models, with support for flexible latent
structures, spline effects, and mortality data applications.
}
\author{
Pedro Menezes de Araújo
}
\keyword{bayesian}
\keyword{clustering}
\keyword{mixture}
\keyword{mortality}
\keyword{package}
