% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_run.R
\name{single_run}
\alias{single_run}
\title{Single run}
\usage{
single_run(
  M,
  w = NULL,
  model_data,
  lambda,
  init_list,
  iters,
  burn_in,
  thin,
  verbose,
  seed = NULL
)
}
\description{
Run MCMC for one single chain
}
