% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_run_temp.R
\name{single_run_temp}
\alias{single_run_temp}
\title{Single run}
\usage{
single_run_temp(
  M,
  w = NULL,
  model_data,
  lambda,
  intercept_penalty,
  dirichlet_param,
  init_list,
  iters,
  burn_in,
  thin,
  verbose,
  add_logpost = TRUE,
  seed = NULL
)
}
\description{
Run MCMC for one single chain
}
